feat: Use long instead of int to represent pipeline ids
Change type of pipeline id in aliases GitLabGetPipelineAsync() and GitLabSetPipelineNameAsync() from int to long to match the type used by NGitLab

+++ src/Cake.GitLab/_Aliases/GitLabAliases.Pipelines.cs
@@ -17,7 +17,7 @@ public static partial class GitLabAliases
     /// <param name="pipelineId">The id of the pipeline to load data for.</param>
     [CakeMethodAlias]
     [CakeAliasCategory("Pipelines")]
-    public static async Task<Pipeline> GitLabGetPipelineAsync(this ICakeContext context, string serverUrl, string accessToken, ProjectId project, int pipelineId) =>
+    public static async Task<Pipeline> GitLabGetPipelineAsync(this ICakeContext context, string serverUrl, string accessToken, ProjectId project, long pipelineId) =>
         await context.GetGitLabProvider().GetPipelineAsync(serverUrl, accessToken, project, pipelineId);
 
     /// <summary>
@@ -31,6 +31,6 @@ public static partial class GitLabAliases
     /// <param name="name">The name to set the pipeline name to.</param>
     [CakeMethodAlias]
     [CakeAliasCategory("Pipelines")]
-    public static async Task GitLabSetPipelineNameAsync(this ICakeContext context, string serverUrl, string accessToken, ProjectId project, int pipelineId, string name) =>
+    public static async Task GitLabSetPipelineNameAsync(this ICakeContext context, string serverUrl, string accessToken, ProjectId project, long pipelineId, string name) =>
         await context.GetGitLabProvider().SetPipelineNameAsync(serverUrl, accessToken, project, pipelineId, name);
 }
